refactor: splitting up calculate-score in different files
