+++ src/engine/define-initial-state.ts
@@ -1,0 +1,10 @@
+import { RunGameData, RunGameState, RunGameStatus } from 'data/interfaces'
+
+export function defineInitialState(run: RunGameData): RunGameState {
+  return {
+    status: RunGameStatus.Planned,
+    teams: Object.keys(run.teams).map(key => {
+      return { team: key, rank: 0, score: 0, bingos: 0, completedChallenges: [] }
+    }),
+  } as RunGameState
+}
