+++ src/engine/types.ts
@@ -1,0 +1,6 @@
+export class EngineError extends Error {
+  constructor(message: string) {
+    super(`[engine] ${message}`)
+    this.name = 'EngineError'
+  }
+}
